requirements.txt
# Django 기본 (API 서버)
Django>=5.0.0
djangorestframework
django-cors-headers
gunicorn

# Notion API 연동
notion-client
requests
requests-oauthlib

# PWA 및 캐싱
django-pwa
redis  # 서버사이드 캐싱

# 엑셀 내보내기 (선택사항)
openpyxl
xlsxwriter
pandas

# 이미지 처리
Pillow
python-magic

# 기타 유용한 라이브러리
django-extensions
django-debug-toolbar
python-decouple
python-dotenv
dateutil

# HTTPS/SSL 지원
pyOpenSSL
certifi

# 개발/테스트용
django-silk  # API 성능 모니터링
coverage     # 테스트 커버리지

# 인증 & 보안
argon2-cffi
django-environ

# 유틸리티
ipython
chardet
arrow

# 시스템 성능 모니터링
psutil

# AI/ML 분석 (간단한 라이브러리)
numpy
scipy
django-widget-tweaks